# test-constructor/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для конструктора тестов",
        "title": "Test Constructor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/intern/tests/selection": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает список доступных тестов для мероприятия с учётом замен",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "Получить список тестов",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список тестов",
                        "schema": {
                            "$ref": "#/definitions/dto.TestSelectionResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Пользователь не записан на мероприятие",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/manager/events": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список всех мероприятий из CRM",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить мероприятия",
                "responses": {
                    "200": {
                        "description": "Список мероприятий",
                        "schema": {
                            "$ref": "#/definitions/dto.EventsListResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создает новую конфигурацию тестирования для мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Создать конфигурацию мероприятия",
                "parameters": [
                    {
                        "description": "Конфигурация",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEventConfigRequest"
                        }
                    },
                    {
                        "description": "Specialization ID (0 = общий тест)",
                        "name": "specialization_id",
                        "in": "body",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Конфигурация создана",
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEventConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/events/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновляет существующую конфигурацию мероприятия",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Обновить конфигурацию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Config ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленная конфигурация",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEventConfigRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Конфигурация обновлена",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEventConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Нет прав",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Не найдена",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/manager/events/{id}/specializations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получает список специализаций конкретного мероприятия из CRM",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Получить специализации мероприятия",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список специализаций",
                        "schema": {
                            "$ref": "#/definitions/dto.EventSpecializationsListResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Мероприятие не найдено",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateEventConfigRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "extra_threshold": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ExtraThresholdRequest"
                    }
                },
                "fail_text": {
                    "type": "string"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "success_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "integer"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateEventConfigResponse": {
            "type": "object",
            "properties": {
                "config_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.EventResponse": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "2024-12-02T00:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Хакатон 2024"
                },
                "specializations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EventSpecializationResponse"
                    }
                },
                "start_date": {
                    "type": "string",
                    "example": "2024-12-01T00:00:00Z"
                },
                "total_tests": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "dto.EventSpecializationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Frontend Developer"
                }
            }
        },
        "dto.EventSpecializationsListResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer",
                    "example": 1
                },
                "specializations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EventSpecializationResponse"
                    }
                }
            }
        },
        "dto.EventsListResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.EventResponse"
                    }
                },
                "total": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "dto.ExtraThresholdRequest": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_threshold": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "integer"
                }
            }
        },
...
        "dto.TestInfo": {
            "type": "object",
            "properties": {
                "attempt_id": {
                    "type": "integer"
                },
                "config_id": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "is_common": {
                    "type": "boolean"
                },
                "is_extra": {
                    "type": "boolean"
                },
                "max_score": {
                    "type": "integer"
                },
                "passed": {
                    "type": "boolean"
                },
                "replaced_test_id": {
                    "type": "integer"
                },
                "replaced_title": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                },
                "status": {
                    "description": "available, locked, in_progress, completed",
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "test_link": {
                    "type": "string"
                },
                "time_limit": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.TestSelectionResponse": {
            "type": "object",
            "properties": {
                "all_completed": {
                    "type": "boolean"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_passed": {
                    "type": "boolean"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "tests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TestInfo"
                    }
                }
            }
        },
...
        "dto.UpdateEventConfigRequest": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "extra_threshold": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ExtraThresholdRequest"
                    }
                },
                "fail_text": {
                    "type": "string"
                },
                "specialization_id": {
                    "type": "integer"
                },
                "success_text": {
                    "type": "string"
                },
                "test_id": {
                    "type": "integer"
                },
                "threshold": {
                    "type": "integer"
                },
                "time_limit": {
                    "type": "integer"
                }
            }
        },
        "dto.UpdateEventConfigResponse": {
            "type": "object",
            "properties": {
                "config_id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите токен в формате: Bearer {ваш_токен}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}